+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.6.3
+
++ Ignore unknown formatters.
+
 ## 0.6.2
 
 + Change regex for email validation (~7x faster).